requirements.txt
phidata
fastapi
uvicorn
groq
packaging
duckduckgo-search
python-dotenv
yfinance
openai